Refactor: Move GPT image logging functions to separate utility file
